ILLUSIONS: DUCKMAN: Implement opcodes 70/71 for pausing/unpausing during the menu
Also change some input functions to return uint istead of byte to allow bigger bitmasks.

+++ engines/illusions/input.h
@@ -67,7 +67,7 @@ public:
 	InputEvent& setBitMask(uint bitMask);
 	InputEvent& addKey(Common::KeyCode key);
 	InputEvent& addMouseButton(int mouseButton);
-	byte handle(Common::KeyCode key, int mouseButton, bool down);
+	uint handle(Common::KeyCode key, int mouseButton, bool down);
 	uint getBitMask() const { return _bitMask; }
 protected:
 	uint _bitMask;
